created utils folder inside compensations folder and created validation file to align with PR 535
